extract get plugin id function for general use
